* commit 'bc223ff78cf3d56e74e9802daaff72b798e0f023': (24 commits)
  Fix to multi-chromosome loads failing dnarun due to invalid duplicated data. Invalid data is a warning, not an error, and will no longer fail a load.
  remove extra semicolon
  Correct errors in package naming of refactored map classes
  change version to 1.4
  Technically, it makes more sense to disallow this as an option, but I feel like this could be useful as a 'neat trick' to reduce storage space usage in real time, so I'm keeping it around, it's just hidden in a command line argument that has to be set.
  Fix to '-C' being invalid VCF data.
  that's not a semicolon!
  Don't delete intermediate files on a successful load if 'keepIntermediateFiles' flag is set.
  re-organize dto namespaces
  add documentation and class constraint to DtoMap
  move exception class to core namespace
  Move remaining classes to the aspected class
  Move DtoOrganization to the aspected class
  Move DtoMapSet to the aspected class
  Move DtoMapManifested to the aspected class
  Move DtoMapExperiment to the aspected class
  Move DtoMapAnalysis to the aspected class
  Change target of Aspect
  Move DtoMapContact to the aspected class
  mark date/user parameters nullable
  ...

+++ gobii-process/src/main/java/org/gobiiproject/gobiiprocess/digester/GobiiFileReader.java
@@ -16,8 +16,8 @@ import org.gobiiproject.gobiiclient.core.gobii.GobiiEnvelopeRestResource;
 import org.gobiiproject.gobiimodel.cvnames.JobProgressStatusType;
 import org.gobiiproject.gobiimodel.config.*;
 import org.gobiiproject.gobiimodel.dto.instructions.extractor.*;
-import org.gobiiproject.gobiimodel.headerlesscontainer.DataSetDTO;
-import org.gobiiproject.gobiimodel.headerlesscontainer.ExtractorInstructionFilesDTO;
+import org.gobiiproject.gobiimodel.dto.entity.auditable.DataSetDTO;
+import org.gobiiproject.gobiimodel.dto.instructions.extractor.ExtractorInstructionFilesDTO;
 import org.gobiiproject.gobiimodel.utils.DateUtils;
 import org.gobiiproject.gobiimodel.utils.FileSystemInterface;
 import org.gobiiproject.gobiiapimodel.payload.HeaderStatusMessage;
@@ -441,10 +441,12 @@ public class GobiiFileReader {
                         loadedData = true;
                     }
 					if(counts.invalidData >0 && !isVariableLengthTable(key)){
-						ErrorLogger.logError("FileReader","Error in table "+key);
+						ErrorLogger.logWarning("FileReader","Invalid data in table "+key);
 					}
 					else{
-						deleteIFLFiles(dstDir,key);
+						if(LoaderGlobalConfigs.getDeleteIntermediateFiles()){
+							deleteIFLFiles(dstDir,key);
+						}
 					}
 
 				}
